one map page, marker styling, dynamic coloring
