+++ server/src/routes/api/v1/markersRouter.js
@@ -1,6 +1,9 @@
 import express from "express"
+import { ValidationError } from "objection"
+import cleanUserInput from "../../../services/cleanUserInput.js"
 import Marker from "../../../models/Marker.js"
 import MarkerSerializer from "../../../serializers/MarkerSerializer.js"
+import upload from "../../../services/photoUpload.js"
 
 const markersRouter = new express.Router()
 
@@ -13,5 +16,27 @@ markersRouter.get("/", async (req, res) => {
     return res.status(500).json({ errors: error })
   }
 })
+
+markersRouter.post("/", upload.single("photo"), async (req, res) => {
+  const body = req.body
+  const cleanBody = cleanUserInput(body)
+
+  const newData = {
+    ...cleanBody,
+    photo: req.file.location,
+    userId: req.user.id
+  }
+
+  try {
+    const newMarker = await Marker.query().insertAndFetch(newData)
+    return res.status(201).json({ newMarker })
+  } catch (error) {
+    console.log(error)
+    if (error instanceof ValidationError) {
+      return res.status(422).json({ errors: error.data })
+    }
+    return res.status(500).json({ errors: error })
+  }
+})  
 
 export default markersRouter
