+++ server/src/routes/clientRouter.js
@@ -3,7 +3,7 @@ import getClientIndexPath from "../config/getClientIndexPath.js";
 
 const router = new express.Router();
 
-const clientRoutes = ["/", "/user-sessions/new", "/users/new", "/mymap", "/discover"];
+const clientRoutes = ["/", "/user-sessions/new", "/users/new", "/map"];
 router.get(clientRoutes, (req, res) => {
   res.sendFile(getClientIndexPath());
 });
